+++ mdook/__init__.py
@@ -2,4 +2,4 @@
 libraries and single documents.
 """
 
-__version__ = "2.0.0"
+__version__ = "2.0.1"